requirements.txt
DateTime==4.3
Django==3.2.4
django-model-utils==4.1.1
djangorestframework==3.12.4
numpy==1.21.1
pip==21.1.2
python-dateutil==2.8.0
requests==2.25.1
urllib3==1.26.5
Pillow==8.3.1